+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* Add a dashboard migration tool for converting WordPress users into guest author profiles by @faisalahammad in https://github.com/Automattic/Co-Authors-Plus/pull/1344
+
 ## [4.1.1] - 2026-06-26
 
 ### Fixed